--- a/python/setup.cfg
+++ b/python/setup.cfg
@@ -8,14 +8,6 @@ url = https://github.com/equinor/oneseismic
 version = 0.2.1
 
 [options]
-packages =
-    oneseismic
-    oneseismic.client
-    oneseismic.internal
-    oneseismic.login
-    oneseismic.scan
-    oneseismic.upload
-
 install_requires =
     azure-storage-blob
     gql
@@ -30,6 +22,8 @@ install_requires =
 
 setup_requires =
     pytest-runner
+    scikit-build
+    pybind11
 
 tests_require =
     hypothesis