There's actually a couple of (somewhat related) changes in this PR, but it brings two major improvements.

## Performance improvements
I measured the time taken by parse+extract _after_ all data was downloaded before and after this PR.

    NEW
        0.001031s
        0.000731s
    OLD
        0.024170s
        0.020067s

Which means that in parsing and extracting alone this brings 20-30x performance improvement.

## Message format
The message format has been reviewed, and the (main) payload is no longer serialized as an array, but rather bin. The messagepack binary type is really just length + blob, and leaves the actual format to the application. In this case we know it's float32 little-endian, which brings down overhead (1 byte per element) and reduces serialization cost. Most of the performance improvements come from this change.

## C++ decoder
I've written a decoder in C++ that should cover the needs of all clients, which should implement an efficient protocol and streaming, and make it (slightly) easier to tune the message format. This is a design choice, since a C++ decoder can be used from virtually all languages. The main driver here is Javascript; we can provide an emscripten lib that JS clients can use to leverage oneseismic directly from the browser.

The C++ decoder brings _some_ performance gains (mostly in predictability), but it is not yet used in a streaming fashion.

## Misc
Some good tests are sorely missing, but right now the format is in significant flux, and a test suite for the decoder needs to be planned and executed well if it is to make any sense. This should probably be prioritised soon.